Setting new format as base
Though this does represent a new file format there are subfolder sets that need to be better cleaned up on.

One notable spot is any river folder has a "new version" folder. I was not sure what to do with this as it was basically a copy of all of the river art and same names.